OMAP:OMAPLFB: Added reference counting for NV12 buffers on display
Added reference count for each time an NV12 buffer is queued to
the display.  It is decremented when the buffer is swapped with
another one (including the same one queued again).

This is a crude solution since SGX is not keeping proper readops
count for displayed buffers.

Using atomic operations to change the reference count as the
incrementing and decrementing are done in different blocks
(graphics and display) under different locks.  This should
still be OK because the incrementing and the checking is in
the graphics context.

Requires corresponding change in DDK.

Change-Id: I361577a3104e002cf9e5c5452fb2ce967545766b
Signed-off-by: Lajos Molnar <molnar@ti.com>