ARM: omap4: wakeupgen: unmask based on node, not affinity
The irq affinity mask is not kept up to date by the ARM irq
code.  In particular, migrate_one_irq is not called when
the irq node does not match the cpu going offline, and
when it is called, it doesn't update the irq affinity
mask before calling the irq chip's set_affinity function.
This causes the use of the affinity mask in the mask
and unmask functions to be unreliable, possibly unmasking
an interrupt on a cpu that is offline.

The gic driver only supports directing an irq to a single
cpu at a time, selected by the node field in the irq_data,
so use that to mask and unmask interrupts in the wakeupgen.

Change-Id: I1da9ed4cb2ebddacebce51f6f2f8fa054181a259
Signed-off-by: Colin Cross <ccross@android.com>