OMAP4: usb: Move VUSB source and pulldown to regulator
The VUSB regulator has additional settings in the MISC2
and CFG_LDO_PD2 registers.  These settings affect the
source of the regulator and the pull down state in off
mode.

This change initializes these values when the regulator
is initialized.  Previously these regulator settings
were configured in the twl6030-usb driver.

Signed-off-by: Adam Hampson <ahampson@sta.samsung.com>