OMAP:OMAPLFB: Use TILER 2D buffer for swap chain
Use a TILER 2D buffer for swap chain instead of linear framebuffer,
so that we can rotate it on a display.

First allocate a secondary 2D area as swapchain buffer (for 3 buffers).
Then override framebuffer information when populating PVRFB information
so that it points to this 2D area.

Use the same number of buffers as if we allocated from FB VRAM, but
limit swap chain to max 3 buffers to minimize TILER container use.
For some reason Android needs at least 3 swap chain buffers.

Change-Id: I1ba39c1108246747d3153d41875a3638bd40671f
Signed-off-by: Lajos Molnar <molnar@ti.com>