Revert "OMAP4: DSS: Fix CIO errors with interleaved commands"
This reverts commit 0af462f8b6ba1c5e8b5d684e4c1679349bec22fd.

That commit breaks dsi_vc_dcs_read in video mode. It also does not
seem good for dsi_vc_dcs_write since it makes that function an alias
for dsi_vc_dcs_write_nosync. Without the response the call does not
block until the data has been received on the other end, which
can cause problems with init sequences that require delays
between commands, and it cannot report errors that are part of
the response.

Change-Id: Id0c32e675b3c2f834575b4b23f296ede5da0bee9
Signed-off-by: Arve Hjønnevåg <arve@android.com>