OMAP4: HSI: Tesla: set to RET after wakeup
Upon HSI cawake, it also wakesup Tesla because of satic dep
set HSI towards DSP. Tesla has to be reset back so that it
can be allowed to idle again.
As per observation, this issue seems to occur with OFF mode enabled.
With RET mode Tesla idles properly. However, this patch keeps the
workaround as is and sets the next state of tesla to RET. Else this causes
unnecessary sleep switch on Tesla in late suspend call causing
it to be ON always.

REVISIT needed:
- When OFF mode is enabled, for setting proper next state.
- For verifying power-on reset of DSP is done correctly.
- For removing magic numbers.

Change-Id: I74cc1528cf501c46d86dec8f2d7e35c9e2b88d49
Signed-off-by: Girish S G <girishsg@ti.com>