rpmsg: res_mgr: add i2c as resource
Add i2c to the list of requestable resources. This will allow us
to keep the i2c clocks enabled for the requested i2c bus. However
Ducati decides which bus the system requires. Since the interrupt
handling for the i2c bus is done in Ducati, it is not required
to prevent MPU from idling - hence we dont hold pm_qos request and
allow Ducati to drive the requests on a need basis.

NOTE: In case of a shared I2C bus which is used by A9 and Ducati,
we may need to use hwspinlock to ensure the devices are used in
a mutually explicit manner. This needs to be done on a subsequent
patch.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Acked-by: Nishanth Menon <nm@ti.com>
Change-Id: Ib5bfeefc80cee161e7de912e50fa161424f0e1cc