OMAP:DSSCOMP: simplify queuing
In preparation for HDMI cloning support, simplify queueing in
DSSCOMP queue.c.

dsscomp_t objects are no longer maintained on queues.  Instead,
compositions waiting to be applied are on the apply workqueue.
Any other compositions (not yet queued to be applied, or
already applied) are not tracked.  Compositions are also
deleted on the RELEASED callback.

Therefore, wait operations are moved to device.c that
maintains the state of compositions queued via ioctls.

Gralloc compositions are tracked in gralloc.c, and they are
on a FIFO queue to guarantee in-order completion.  This also
allows such compositions to depend on multiple DSS compositions
queued to different displays.

Change-Id: I60de75bb0ee6024fb072f939cdaf2845bbe68bf9
Signed-off-by: Lajos Molnar <molnar@ti.com>