ARM: omap4: cpuidle: allow all states when both cpus are active
Add a power sequencing step to omap4_enter_idle that
waits for the other cpu to be ready, and then go to the
higher of the two requested idle states.  When both cpus
are ready, sequence the power off so that it follows the
same order as when hotplug is used: cpu1 off, then cpu0 off,
and when powering on, cpu0 on, then cpu1 on.

In order to minimize power when waiting for the other cpu,
the waiting cpu should be in wfi.  omap4_idle_wait waits
in wfi for either an interrupt to abort suspend, or an
event from the other cpu (emulated with an unused interrupt
triggered directly into the GIC).

Change-Id: I8a8e49529c674aeda1584e851fe9eb11cde8fc09
Signed-off-by: Colin Cross <ccross@android.com>