OMAP:OMAPLFB/OMAPDSS: Added support for cloning
Changed gralloc_queue to use a struct that specifies all displays

Cloned buffers are not referenced in multiple slots of the buffers
array for Post2. Instead cloned overlays refer to the index of
the buffer to be cloned.

Overlays that were specified in prior compositions, but are no
longer specified will be blanked (including overlays on managers
that are no longer specified.)

Also changed to return and error instead of BUG() if receiving
invalid struct in ProcessFlipV2.

Change-Id: Icbed3e43d4871781ffc5124fae77fb813f8b8dca
Signed-off-by: Lajos Molnar <molnar@ti.com>