ASoC: ABE DSP: Share firmware strategy for built-in and modules
ABE firmware was stored locally in ABE HAL source code for built-in
support and a new firmware was created for modules which contains
the ABE firmware + equalizer configuration and coefficients.
In order to reuse the code for parsing the new firmware, original
firmware (pure ABE firmware) is converted into new firmware scheme.

Change-Id: I67346a0ae98e5bec238e12562efe571f4c42b84d
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Sebastien Guiriec <s-guiriec@ti.com>