OMAP4: PM: powerdomain: Dont write to readonly/reserved register
Some of the powerst control/logic_st_ctrl registers such as
CAM, EMU, ABE, etc are reserved registers, we should'nt be
writing to them in the first place. The ALWAYS_ON power domains
(wakeup, those of core, DSP, MPU), STD_EFUSE are not under
software control, hence for the purposes of this patch, considered as
not modifiable by software.

Change-Id: I501826e9217583f7562a164703c0c65d3eed0b00
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>