OMAP: 4430sdp/panda: ensure correct order of DMM/Tiler device creation
The power management system uses the order of device creation to
determine the order of suspend and resume function calls.  A number of
existing drivers are dependent on the Tiler resuming before they do.

The dmm and Tiler devices need to be created in the board init functions
so that they will be earlier in the device tree.

Change-Id: I0598b21ffa8a96f0de2f29a4fc18f75667d664e2
Signed-off-by: Andy Gross <andy.gross@ti.com>
Signed-off-by: Iliyan Malchev <malchev@google.com>