TUNA: Add hdmi device to board-tuna-display
Add hdmi device to board config and setup the mux

Use GPIO 63 as HDMI HPD line.

HDMI vref is made as a supply and requested by the driver.
On omap4460 the default internal pullups are turned on.
These are strong pull ups. In tuna we have external pull ups
So we have to disable the internal pullup to make sure the
DDC lines are operational.

Change-Id: I6867e94fa9848d74428d1b5410130607ce3654f0
Signed-off-by: Dandawate Saket <dsaket@ti.com>