ASoC: McPDM: Combine channels management for UL and DL
Due to HW bug on McPDM CTRL register both UL and DL channels need
to be configure at the same time. If the channels are not
configured at the same time, the side effects can be muting in
uplink or loud noise in downlink during stress test transition
between uplink and downlink. Each time McPDM is started both UL
and DL channels are enabled to work around the problem.

Change-Id: Ie5000149df1324e3fdd2a07836d43e15412a2372
Signed-off-by: Sebastien Guiriec <s-guiriec@ti.com>
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>