Revert "omap4: clockdomain: enable forced wakeup and sleep for ipu & dsp"
This reverts commit 12256aea599a56e2835c3f8230eeae7c22a633f7.

The IPU and DSP need to be placed in the SW_WKUP clock state
transition control mode while their requisite module clocks are being
enabled. Once the module clocks are enabled and the processor(s)
are taken out of reset, the clock state transition control mode may
be changed to HW_AUTO which allows sleep transitions to be
managed automatically by the hardware, according to the processor
and slave component states.

The functions clkdm_wakeup( ) and clkdm_allow_idle( ) set the
SW_WKUP and HW_AUTO clock state transition control values
respectively.

Setting both HWSUP and SWSUP flags allows these modules to
activate both clock state transition control modes as required
when the one of the above clkdm functions is called.

The remoteproc driver is responsible for calling those API's
in order to allow the proper initialization and functionality.

Conflicts:

	arch/arm/mach-omap2/clockdomains44xx_data.c

Change-Id: I647c5f7a11b8d692a101f329424dfb3ef04d2dee
Cc: Benoit Cousson <b-cousson@ti.com>
Cc: Rajendra Nayak <rnayak@ti.com>
Cc: Suman Anna <s-anna@ti.com>
Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>