remoteproc: make rproc_set_secure a blocking call
rproc_set_secure call is now made into a blocking call.
It will be blocked till the cleanup is complete and a
restart is being attempted.

THe function will return the status based on whether the
intended operation (secure reload or normal reload) has
succeeded. Note that this doesn't return the status of
the rpmsg channels creation.

Change-Id: Ia0a02d261ba67f06235cd6bd5826a47319592956
Signed-off-by: Suman Anna <s-anna@ti.com>