OMAP: Add debugfs node to show the summary of all clocks
From: Jon Hunter <jon-hunter@ti.com>

Add a debugfs node called "summary" to /sys/kernel/debug/clock/
that displays a quick summary of all clocks registered in the
"clocks" structure. The format of the output from this node is:

<clock-name> <parent-name> <rate> <usecount>

This debugfs node was very helpful for taking a quick snapshot of
the linux clock tree for OMAP and ensuring clock frequencies
calculated by the kernel were indeed correct. This patch helped
uncover some bugs in the linux clock tree for OMAP4.

Change-Id: Ife6a298ee5f2532f0154e345ef311e15b266a21d
Signed-off-by: Jon Hunter <jon-hunter@ti.com>