omap: dmtimer: Creation of new internal timer APIs
This patch is to create new internal timer APIs for
__timer_enable and __timer_disable. This is done to
ensure that when we implement spin locks we do not
result in problems since current implementation calls
exported APIs from within the internal implementation of this
DMTimer. This is not right when spin_lock is introduced. Hence
the change.

There is no functional change with this Patch only new internal
APIs created as described above.

Change-Id: I0d67a70b3094e2ea793d2e6d213eaf9049f69dde
Signed-off-by: Pradeep Bhat <pradeep.bhat@ti.com>