ASoC: DSP: Use per FE/BE state
Handle states of FEs and BEs in their DSP runtime data, as
opposite to handling it in DSP link. This allows to have a
single reference of FEs/BEs state which is key when doing
state transitions.

New states are defined for FEs/BEs which are a one-to-one
map with PCM states. The DSP link states will represent
only if a particular link is active or ready for pruning.

Change-Id: Icdd5ea33efb1277ad341693d1b3b0d99d2d1fa3d
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>