OMAP4: L2: Don't set POR chicken bit 25.
As per hardware team clarification, the bit 25 in L2X0 POR is
not a copy of Cache replacement policy bit from AUXCTRL.
Its a chicken bit and SBZ, so don't set it.

In AUXCTRL, default value of bit 25 is 1 and since we do a read
modify write, there is no need to set it explicitly.

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Rajeev Kulkarni <rajeevk@ti.com>
---