OMAP: EMIF: diable and clear irqs on shutdown and boot
Introduce a shutdown hook to be called before warm
reboot and before HWMOD initializes.

Since EMIF does not have software reset mechanism,
sometimes during a warm reset scenario, a irq may be pending
before the driver initializes, which can cause the
module to be left in transition and HWMOD initialization
to fail.

This patch disables and clears the IRQ's on both the
EMIF modules on shutdown and as early as possible at bootup.

Change-Id: Ibabbeea8f2bf5ee8e76c2d1d2a8abc6a420eef17
Signed-off-by: Axel Haslam <axelhaslam@ti.com>