OMAP4: PM: Throttle CPU with delay of 10 sec
Once the system reaches T_hot CPU is throttled. The next throttle of
the CPU should be delayed by 10 secs, this is the time taken to cool
down from 95C to ~80C. The idea is not to throttle aggressively as this
will have impact on system performance. When CPU is loaded heavily the
thermal level is maintained below 95C which is the safest operating range
and at the same time keeps CPU running at max available frequency.

Change-Id: I0b8dca25c2efbc2bbb6f7096ddce7ddca7a19d89
Signed-off-by: Girish S G <girishsg@ti.com>