rpmsg: serialize transmissions and add proper barriers
The rpmsg_send_offchannel_raw delivers messages to the remote
processor by getting a vring buffer, copying the necessary
payload and making it available to the remote processor. These
three steps need to be atomic to ensure serialized delivery of
messages to the remote processor.

Memory barriers have also been added to the virtio_rpmsg_bus
driver, since this driver uses the virtio framework across
different processor sub-systems and the internal barriers in
virtio driver are better suited to work within a SMP
environment.

TODO:
The wait logic when no buffer is available needs to be moved
out from this function. This code path currently is never
reached.

Change-Id: I91685ffbccc6769f60e77a2a7106c3a9daa55b69
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Iliyan Malchev <malchev@google.com>