mfd: twl6030: add support for power resource defaults
TWL6030 resources are designed to be used with multiple processors
we should ideally be able to describe the way the board is wired up.

The default configuration which is almost universally used is to
connect TWL6030 only to Application Processor(example: OMAP4430/4460)
and the request lines for 2(a,b,c) and 3 are completely unused.

The register defaults may not reflect this ground reality, so map
it back to the way we want it to look so not to depend on any errors
by bootloaders/wrong pulls.

Change-Id: I2777715b9f63e35f317453df04288ceba45e5e55
Signed-off-by: Nishanth Menon <nm@ti.com>