ASoC: omap-mcasp: Relax allowed sample rates check
McASP playback rates are determined by the frequency of its
FCLK and the clock dividers (CLKXDIV and HCLKXDIV). Given
a particular FCLK rate, there might be sampling frequencies
which cannot be achieved with the McASP clock dividers within
an error margin, those will be rejected.

Sampling frequencies within the allowed PPM margin will be
accepted, so clock dividers computation is relaxed so that exact
matches is not a constraint anymore.

Change-Id: I82ca969960776461b7eab3d0c0f1bf573b042288
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>