OMAP4: PM: restore GPIO clocks during wake from failed OFF
While attempting OFF, we switch off clks for the GPIO blocks. when
device does achieve OFF, we restore effectively and bring configuration
back to order. However, when an attempt to OFF fails, context is not
lost, but the clks to the GPIO blocks are cut. This results in
accesses to GPIO resulting in aborts.

Restore context and configuration based on the fact that we planned
to go to OFF (device_next_state) instead of device actually achieving
OFF or not (prev_state), on resume.

Change-Id: I601c913fefea2ab40ad1e50d7a648d6d2cf34f22
Signed-off-by: Wenbiao Wang <wwang@ti.com>