OMAP2+: PM: provide mechanism to describe overall behavior of osc and PMIC.
We currently have mechanisms in place to describe the PMIC per
rail, however we also need to configure the system for situations
such as OFF mode, where, oscillator switch off and on time, and
similar durations for PMIC also tends to play a major factor.
Introduce a few apis to OMAP2's pm framework to use these. OMAP1
does'nt seem to need this at the moment, hence not a OMAP generic
framework. The set functions are meant to be used by initialization
code.

The OMAP specific implementation would need to use this ofcourse.

Change-Id: I374c638477ee5625a38a7ce49e755c2239e52cc8
Signed-off-by: Nishanth Menon <nm@ti.com>